Add ClassStateMapping
